organization/questionnaire.md: Replace 'student disappears' with 'disappearing students'
With 'the student disappears', it's not clear which is the main noun.
While the present participle 'disappearing' [1,2] can function as a
noun, it's less ambiguous ('student' is always a noun).  This argument
doesn't sound entirely convincing to me, but the new form also echos
the question, so it can't be that wrong ;).

[1]: http://en.wiktionary.org/wiki/disappearing
[2]: http://en.wiktionary.org/wiki/present_participle